+++ src/components/wallet/Wallet.tsx
@@ -1,27 +1,22 @@
-import React, { FC, useMemo } from 'react';
-import { useWallet } from 'use-wallet';
+/* eslint-disable no-console */
+import React, { FC } from 'react';
 import styled from 'styled-components';
-
-import {
-  useIsWalletConnecting,
-  useResetWallet,
-  useConnectWallet,
-  useWalletState,
-  useWalletConnector,
-} from '../../context/AppProvider';
-import { AVAILABLE_CONNECTORS, CONNECTORS } from '../../web3/connectors';
 import { Button } from '../core/Button';
 import { H3 } from '../core/Typography';
 import { Address } from '../core/Address';
-import { ActivitySpinner } from '../core/ActivitySpinner';
 import { FlexRow } from '../core/Containers';
-import { FontSize } from '../../theme';
 import { PageHeader } from '../pages/PageHeader';
 import { ReactComponent as AccountIcon } from '../icons/circle/account.svg';
 import { Balances } from './Balances';
 import { HistoricTransactions } from './HistoricTransactions';
 import { Transactions } from './Transactions';
-import { Connector } from '../../types';
+import { useCloseAccount } from '../../context/AppProvider';
+import {
+  useWalletAddress,
+  useConnected,
+  useWallet,
+  useReset,
+} from '../../context/OnboardProvider';
 
 const Container = styled.div`
   flex: 1;
@@ -59,90 +54,25 @@ const AddressGroup = styled.div`
   justify-content: space-between;
 `;
 
-const Error = styled.div`
-  color: ${({ theme }) => theme.color.red};
-  font-weight: bold;
-  margin: ${({ theme }) => theme.spacing.m} 0;
-`;
-
-const ConnectorsList = styled.div`
-  padding: 0;
-  margin: 0;
-  list-style: none;
-  display: grid;
-  width: 100%;
-
-  grid-template-columns: repeat(auto-fill, minmax(120px, 1fr));
-  grid-column-gap: ${({ theme }) => theme.spacing.m};
-`;
-
-const ConnectorIcon = styled.div`
-  display: flex;
-  flex-grow: 1;
-  align-items: center;
-  justify-content: center;
-  height: 96px;
-
-  > * {
-    width: 50%;
-    height: auto;
-  }
-`;
-
-const ConnectorLabel = styled.div``;
-
-const ConnectorButton = styled(Button)`
-  margin-bottom: 16px;
-  padding: 8px;
-  background: rgba(255, 255, 255, 0.1);
-  border: none;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  font-size: ${FontSize.s};
-`;
-
-const Connecting = styled.div`
-  width: 120px;
-`;
-
-const Disconnected: FC<{}> = () => {
-  const connectWallet = useConnectWallet();
-  const list: Connector[] = useMemo(
-    () => CONNECTORS.filter(({ id }) => !!AVAILABLE_CONNECTORS[id]),
-    [],
-  );
-
-  return (
-    <ConnectorsList>
-      {list.map(({ id, subType, label, icon: Icon }) => (
-        <ConnectorButton
-          key={subType ?? id}
-          type="button"
-          onClick={() => connectWallet(id, subType)}
-          inverted
-        >
-          <ConnectorIcon>{Icon ? <Icon /> : null}</ConnectorIcon>
-          <ConnectorLabel>{label}</ConnectorLabel>
-        </ConnectorButton>
-      ))}
-    </ConnectorsList>
-  );
-};
-
 const Connected: FC<{ walletLabel: string; account: string }> = ({
   walletLabel,
   account,
 }) => {
-  const resetWallet = useResetWallet();
+  const reset = useReset();
+  const closeWallet = useCloseAccount();
+  const onReset = (): void => {
+    if (reset) {
+      reset();
+      closeWallet();
+    }
+  };
   return (
     <Rows>
       <Row>
         <H3>Connected with {walletLabel}</H3>
         <AddressGroup>
           <Address address={account} type="account" copyable />
-          <DisconnectButton type="button" onClick={resetWallet}>
+          <DisconnectButton type="button" onClick={onReset}>
             Disconnect
           </DisconnectButton>
         </AddressGroup>
@@ -164,11 +94,9 @@ const Connected: FC<{ walletLabel: string; account: string }> = ({
 };
 
 export const Wallet: FC<{}> = () => {
-  const { error } = useWalletState();
-  const connecting = useIsWalletConnecting();
-  const { status, account } = useWallet();
-  const connected = status === 'connected';
-  const wallet = useWalletConnector();
+  const address = useWalletAddress();
+  const connected = useConnected();
+  const wallet = useWallet();
 
   return (
     <Container>
@@ -177,24 +105,17 @@ export const Wallet: FC<{}> = () => {
           icon={<StyledAccountIcon />}
           title="Account"
           subtitle={
-            connected
+            connected && address
               ? `Connected`
-              : connecting && wallet
-              ? `Connecting to ${wallet.label} wallet`
+              : !connected && wallet && address !== undefined
+              ? `Connecting to ${wallet.name as string} wallet`
               : 'Connect wallet'
           }
         />
-        {error ? <Error>{error}</Error> : null}
         <FlexRow>
           {/* FIXME problem when first connecting; updates on unmounted components */}
-          {connected && account && wallet ? (
-            <Connected walletLabel={wallet.label} account={account} />
-          ) : connecting ? (
-            <Connecting>
-              <ActivitySpinner pending />
-            </Connecting>
-          ) : (
-            <Disconnected />
+          {connected && address && wallet && (
+            <Connected walletLabel={wallet.name as string} account={address} />
           )}
         </FlexRow>
       </div>